Move CTCP handling to handler
